LocalFileSystemProvider now starts at Drive View.
IProvider.InitialPath is added to the interface and each provider can implement its root location.  For LocalFileSystemProvider that allows for drives to be explored.  Two Tests are failing in the ProviderViewModelTests:
ShouldBeAbleToOpenSelectedPath
ShouldDisplayCurrentPathProperly
These tests no longer have the same logic as the seperator is no longer used in the LocalFileSystemProvider.  We need to discuss how to handle this change.