Refactor to InitialPath
I null coalesced InitialPath on the CurrentPath property.  This prevents CurrentPath from being null Unless InitialPath is null.  I removed the Elvis Operator and the null Coalescing that occurs in the Reactive objects, If CurrentPath is ever null that is an implementation problem and should now throw.  Implementing IProvider.InitialPath with null is a coding problem that should be resolved not protected against.